requirements.txt
Pillow
opencv-python
progressbar-simple